MVVM is great again!
