+++ Contacts/View/Controls/ContactControl.xaml.cs
@@ -1,18 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
+﻿using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using View.Model;
 
 namespace View.Controls
 {
@@ -35,7 +23,7 @@ namespace View.Controls
         {
             var textBox = sender as TextBox;
             var newText = textBox.Text.Insert(textBox.CaretIndex, e.Text);
-            if (!Contact.PhoneMask.IsMatch(newText))
+            if (!Model.Contact.PhoneMask.IsMatch(newText))
             {
                 e.Handled = true;
             }
@@ -55,7 +43,7 @@ namespace View.Controls
             }
 
             var text = (string)e.DataObject.GetData(typeof(string));
-            if (!Contact.PhoneRegex.IsMatch(text))
+            if (!Model.Contact.PhoneRegex.IsMatch(text))
             {
                 e.CancelCommand();
             }
